Add "inner role" ACLs to debugger results
A query like "user:50" would be incomplete because it was looking only
for system- and queue-level rights. If user 50 were granted rights by
being assigned role membership on a ticket (membership in an "inner
role"), then such results were hidden from the straightforward lookups
of the ACL table this extension had done before this commit.

The reason is that when a user is a member of the AdminCc group on a
ticket, they are _not_ automatically included as members of the AdminCc
group on the queue, or on the system. Presumably, if they were, then
they would have enjoy their AdminCc rights on a global level.

The fix, then, is to include ticket roles in the ACL query. To do so is
not easy, as we have to do a large join across these tables:

    - ACL
    - Groups (queue and/or system role groups, which is where ACLs are assigned)
    - Tickets (to bridge between queue role groups and ticket role groups)
    - Groups (ticket role groups)
    - CachedGroupMembers (principals in ticket role groups)

Though it may be possible to build some monster query that encompasses both
cases (direct privilege granting and ticket role privilege granting),
instead what we do is perform this new "inner role" query using raw SQL
(though taking care to avoid injection attacks), and then pulling out
the IDs. A counterfeit join to be sure, but it does mean we have a
better chance to stash away which ticket IDs, and how many tickets, a
principal has ticket role rights on. The following commits expose that
information in the debugger UI.

We also apply the same improvements for assets and catalogs. Extending
%ParentMap should allow additional objects to participate.