Display inner role information in debugger UI
If we're rendering a result row because it was pulled in from the "inner
role" query (see previous commit) then instead of displaying, say:

    AdminCc
    Queue Role

Then with this commit we can now be more accurate and helpful, instead
displaying:

    AdminCc
    Ticket #3 Role (+4 others)

This is more accurate because the user is _not_ an AdminCc on the queue;
it just appears that way due to ACLEquivalenceObjects. The user is
AdminCc on only 5 tickets.