Linkify "(+x others)" with a TicketSQL search when applicable
If a user is a member of AdminCc on several tickets, the rights debugger
shows only the first, as well as a count of how many others there are.
This is appropriate for such a small UI (listing all the tickets by ID
would be completely bonkers).

You can easily search for which tickets this user is an AdminCc member,
so let's do so. The resulting looks like

    Queue = 'General'
    AND AdminCc.Name = 'user'

If AdminCc rights are granted on a system level, then such ACLs
shouldn't be limited to any queue, so the query is simply:

    AdminCc.Name = 'user'